requirements.txt
Werkzeug==1.0.1
opencv_python==4.3.0.36
Flask==1.1.2
numpy==1.19.1
Pillow==7.2.0
gunicorn==20.0.4